% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetEncodedValueByTraitId.R
\name{dbGetEncodedValueByTraitId}
\alias{dbGetEncodedValueByTraitId}
\title{Get encoded values by trait ID}
\usage{
dbGetEncodedValueByTraitId(db, trait_ids)
}
\arguments{
\item{db}{A database connection}

\item{trait_ids}{a vector of trait_ids}
}
\description{
Get encoded values by trait ID
}

