% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetTableByAccession.R
\name{dbGetTableByAccession}
\alias{dbGetTableByAccession}
\title{Returns a list of complete tables by pht accession}
\usage{
dbGetTableByAccession(phts, db)
}
\arguments{
\item{phts}{A vector of pht accessions}

\item{db}{A connection to topmed_pheno}
}
\description{
Returns a list of complete tables by pht accession
}

