% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetTraitInfoByVarName.R
\name{dbGetTraitInfoByVarName}
\alias{dbGetTraitInfoByVarName}
\title{Get trait info from a table by pht number and trait names}
\usage{
dbGetTraitInfoByVarName(pht, var_names, db)
}
\arguments{
\item{pht}{A dbGaP accession for a table, as an integer}

\item{var_names}{A vector of trait names}

\item{db}{A database connection}
}
\description{
Get trait info from a table by pht number and trait names
}
