% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetTraitInfoByAccession.R
\name{dbGetTraitInfoByAccession}
\alias{dbGetTraitInfoByAccession}
\title{Get trait info by dbGaP phv number}
\usage{
dbGetTraitInfoByAccession(db, dbgap_trait_accessions)
}
\arguments{
\item{db}{A database connection}

\item{dbgap_trait_accessions}{A vector of dbGaP variable accessions, as integers or strings with a leading "phv"}
}
\description{
Get trait info by dbGaP phv number
}

